optimizations for gopherjs
* exchange class list interface for class query/cursor, avoids a very
expensive copy
* remove len(map) where possible — gopherjs implements this with
Object.keys()
* remove Id().String() conversions in favor of string(Id) where possible
* change many things to pointer receivers  — gopherjs clones the object
inside the function, even for non-mutating access: ex. a member field
return.
* float32 -> float64 ( questionable, but recommended )

+++ appengine/datastore/datastore_test.go
@@ -18,7 +18,7 @@ import (
 // https://golang.org/pkg/encoding/binary/#Write
 
 func TestStoreNums(t *testing.T) {
-	vals := []float32{-1.2, 3.1, 4.2}
+	vals := []float64{-1.2, 3.1, 4.2}
 	stream := new(bytes.Buffer)
 	require.NoError(t, writeNums(stream, vals))
 	b := stream.Bytes()
@@ -58,7 +58,7 @@ func TestStoreCoding(t *testing.T) {
 		val   interface{}
 	}
 	xs := []X{{
-		meta.NumProperty, float32(2.3),
+		meta.NumProperty, float64(2.3),
 	}, {
 		meta.TextProperty, "text",
 	}, {
@@ -66,7 +66,7 @@ func TestStoreCoding(t *testing.T) {
 	}, {
 		meta.ObjectProperty, ident.MakeId("object"),
 	}, {
-		meta.NumProperty | meta.ArrayProperty, []float32{1, 2, 3},
+		meta.NumProperty | meta.ArrayProperty, []float64{1, 2, 3},
 	}, {
 		meta.TextProperty | meta.ArrayProperty, []string{"a", "b", "c"},
 	}, {
